Moving functions to utils.py and adding tests
